Updated PortfolioSLO Error Handling
Implemented further error handling using sanitiation functions in update, get, delete and insert. Fixed redundant function calls and return values

+++ src/main/controllers/PortfolioSLOController.js
@@ -5,30 +5,43 @@ class PortfolioSLOController {
     // Generate query payload
     generateCoursePayload(portfolio_id, slo_id) {
         return {
-            portfolio_id: parseInt(portfolio_id),
-            slo_id: parseInt(slo_id)
+            portfolio_id: Janitor.parseIntHandler(portfolio_id),
+            slo_id: Janitor.parseIntHandler(slo_id)
         }
     }
 
     // Get a course
-    async getByAttributes(portfolio_id, slo_id) {    
-        return await PortfolioSLO
-            .query()
-            .where('portfolio_id', parseInt(portfolio_id))
-            .where('slo_id', parseInt(slo_id))
+    async getByAttributes(portfolio_id, slo_id) {
+        var portid = this.sanitizePortfolioId(portfolio_id)
+        var slo_id = this.sanitizeSLOId(slo_id)  
+        if (portid == null || slo_id == null) {
+            return null
+        }  
+        else {
+            return await PortfolioSLO
+                .query()
+                .where('portfolio_id', parseInt(portfolio_id))
+                .where('slo_id', parseInt(slo_id))
+        }
     }
 
     async getById(id) {
-        return await PortfolioSLO
-            .query()
-            .findById(id)
+        var sanitized = Janitor.sanitizeInt(id)
+        if (sanitized == null || sanitized <= 0) {
+            return null
+        }
+        else {
+            return await PortfolioSLO
+                .query()
+                .findById(id)
+        }
     }
 
     // Insert a course
     async insert(portfolio_id, slo_id) {
         var portid = this.sanitizePortfolioId(portfolio_id)  
         var sid = this.sanitizeSLOId(slo_id)   
-        if (portid == 0 || sid == 0) {
+        if (portid == null || sid == null) {
             return null
         }
         else {
@@ -40,60 +53,85 @@ class PortfolioSLOController {
 
     // Update a course
     async updateById(id, portfolio_id, slo_id) {
-        return await PortfolioSLO
-            .query()
-            .patchAndFetchById(id, this.generateCoursePayload(portfolio_id, slo_id))
+        var portid = this.sanitizePortfolioId(portfolio_id)
+        var sid = this.sanitizeSLOId(slo_id)
+        var checkid = Janitor.sanitizeInt(id)
+        if (checkid == null || portid == null || sid == null) {
+            return null
+        }
+        else {
+            return await PortfolioSLO
+                .query()
+                .patchAndFetchById(id, this.generateCoursePayload(portfolio_id, slo_id))
+        }
     }    
 
     // Delete a course
     async deleteByAttributes(portfolio_id, slo_id) {
-        const numDeleted = await PortfolioSLO
-            .query()
-            .delete()
-            .where('portfolio_id', parseInt(portfolio_id))
-            .where('slo_id', parseInt(slo_id))
-        
+        var portid = this.sanitizePortfolioId(portfolio_id)
+        var sid = this.sanitizeSLOId(slo_id)
+        if (portid == null || sid == null) {
+            return null
+        }
+        else {
+            const numDeleted = await PortfolioSLO
+                .query()
+                .delete()
+                .where('portfolio_id', parseInt(portfolio_id))
+                .where('slo_id', parseInt(slo_id))
+
+            return numDeleted > 0
+        }
         // Return true if a course is successfully deleted
-        return numDeleted > 0
+        
     }
 
     async deleteById(id) {
-        const numDeleted = await PortfolioSLO
-            .query()
-            .deleteById(id)
+        var sanitized = Janitor.sanitizeInt(id)
+        if (sanitized == null || sanitized <= 0) {
+            return null
+        }
+        else {
+            const numDeleted = await PortfolioSLO
+                .query()
+                .deleteById(id)
 
+            return numDeleted > 0
+        }
         // Return true if a course is successfully deleted
-        return numDeleted > 0
+    
     }
 
     // sanitize portfolio_id attribute
     sanitizePortfolioId(portfolio_id) {
-        var MAXINT = Math.pow(2,31)         // variable to represent maximum value for a SQL int
+        var MAXINT = Math.pow(2,31)-1         // variable to represent maximum value for a SQL int
 
         // method to check for non-empty input and valid integer
         var san_status
-        if (Janitor.sanitizeInt(portfolio_id) == null || Janitor.sanitizeInt(portfolio_id) <= 0 || Janitor.sanitizeInt(portfolio_id) > MAXINT) {
-            san_status = 0
+        var sanitized = Janitor.sanitizeInt(portfolio_id)
+        if (sanitized == null || sanitized <= 0 || sanitized > MAXINT) {
+            san_status = null
         }
         else {
-            san_status = Janitor.sanitizeInt(portfolio_id)
+            san_status = sanitized
         }
-        return san_status       // 0 means invalid input, 1 means sanitation is successful
+        return san_status       // null means invalid input, else means sanitation is successful
     }
 
     // sanitize slo_id attribute
     sanitizeSLOId(slo_id) {
-        var MAXINT = Math.pow(2,31)         // variable to represent maximum value for a SQL int
+        var MAXINT = Math.pow(2,31)-1         // variable to represent maximum value for a SQL int
         
         // method to check for non-empty input and valid integer
         var san_status
-        if (Janitor.sanitizeInt(slo_id) == null || Janitor.sanitizeInt(slo_id) <= 0 || Janitor.sanitizeInt(slo_id) > MAXINT) {
-            san_status = 0
+        var sanitized = Janitor.sanitizeInt(slo_id)
+        if (sanitized == null || sanitized <= 0 || sanitized > MAXINT) {
+            san_status = null
         }
         else {
-            san_status = Janitor.sanitizeInt(slo_id)
+            san_status = sanitized
         }
-        return san_status       // 0 means invalid input, 1 means sanitation is successful
+        return san_status       // null means invalid input, else means sanitation is successful
     }
 }
